+++ test/mri/excludes/TestEnumerable.rb
@@ -18,5 +18,4 @@ exclude :test_reject, "ArgumentError: wrong number of arguments (given 1, expect
 exclude :test_select, "ArgumentError: wrong number of arguments (given 1, expected 2)"
 exclude :test_sort_by, "ArgumentError: wrong number of arguments (given 1, expected 2)"
 exclude :test_sum, "NameError: uninitialized constant TestEnumerable::SimpleDelegator"
-exclude :test_take_while, "ArgumentError: wrong number of arguments (given 1, expected 2)"
 exclude :test_zip, "ArgumentError: wrong number of arguments (given 1, expected 2)"